BIG DATA AND ANALYTICS
WHAT IS BIG DATA • As Gartner defines it – “Big Data are high volume, high velocity, or high-variety information assets that require new forms of processing to enable enhanced decision making, insight discovery, and process optimization.” • The term ‘big data’ is self-explanatory − a collection of huge data sets that normal computing techniques cannot process. • The term not only refers to the data, but also to the various frameworks, tools, and techniques involved. • Technological advancement and the advent of new channels of communication (like social networking) and new, stronger devices have presented a challenge to industry players in the sense that they have to find other ways to handle the data. • Big data is an all-inclusive term, representing the enormous volume of complex data sets that companies and governments generate in the present-day digital environment. • Big data, typically measured in petabytes or terabytes, materializes from three major sources— transactional data, machine data, and social data